tests/wpt/tests/editing/other/merge-span-with-style-after-pressing-enter-followed-by-backspace-in-contenteditable-div.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<title>Merge Span with style after pressing enter followed by backspace in contenteditable</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
</head>
<body>
<div contenteditable><h1><span style="background-color:red;">Backspace</span></h1></div>
<script>
"use strict";

const kBackspaceKey = "\uE003";
const kEnterKey = "\uE007";

function sendBackspaceKey() {
  return new test_driver.Actions()
    .keyDown(kBackspaceKey)
    .keyUp(kBackspaceKey)
    .send();
}

function sendEnterKey() {
  return new test_driver.Actions()
    .keyDown(kEnterKey)
    .keyUp(kEnterKey)
    .send();
}

promise_test(async () => {
  const editableDiv = document.querySelector("div[contenteditable]");
  const spaceSpan = editableDiv.querySelector('span');
  await new test_driver.click(document.querySelector('span'));
  const range = document.createRange();
  const selection = window.getSelection();
  const textNode = spaceSpan.firstChild;
  range.setStart(textNode, 4);
  range.setEnd(textNode, 4);
  selection.removeAllRanges();
  selection.addRange(range);
  await sendEnterKey();
  await sendBackspaceKey();
  assert_equals(
    editableDiv.innerHTML,
    "<h1><span style=\"background-color:red;\">Back</span><span style=\"background-color: red;\">space</span></h1>",
    "Style is not preserved for the span after pressing backspace in contenteditable"
  );
}, "waiting for command to execute");
</script>
</body>
</html>
